Refactored: Fixtured requests helper class into the gem library
