+++ lib/butterfli-rails.rb
@@ -9,3 +9,6 @@ require 'active_support/dependencies'
 
 # Load files
 require 'butterfli/rails'
+
+# Development dependency:
+# require 'butterfli/rails/test'
